ENH: move methods from 'utils/events.py' to 'Events.py' - 'get_event_filename' and 'get_event_text'.
