+++ scripts/utils/__init__.py
@@ -3,8 +3,6 @@
 
 from . import digits
 from . digits import *
-from . import events
-from . events import *
 from . import logger
 from . logger import *
 from . import photometry
@@ -16,7 +14,6 @@ from . tq_funcs import *
 
 __all__ = []
 __all__.extend(digits.__all__)
-__all__.extend(events.__all__)
 __all__.extend(logger.__all__)
 __all__.extend(photometry.__all__)
 __all__.extend(repos.__all__)
